Moving simulate_match, look_ahead and limited_sim.
